DynamicText: use the raw expression value rather than NaN in the case of objects, which is what Blocks do
The reason Blocks use NaN as a condition return value for objects is that they can't neccessarily tell if an object is changed if its inner properties have changed but the object is the same. This doesn't seem neccessary in the case of DynamicText, since primarily we are rendering strings, and alternatively we are rendering templates. Templates are not intendend to be manipulated dynamically and cause re-rendering, unlike data objects that might be the value of an expresion in a block.